+++ src/main/java/io/travel/model/DayItinerary.java
@@ -3,4 +3,4 @@ package io.travel.model;
 import java.time.LocalDate;
 import java.util.List;
 
-public record DayItinerary(int dayNumber, LocalDate date, List<ItineraryItem> items, double estimatedTotal) {}
+public record DayItinerary(int dayNumber, LocalDate date, List<ItineraryItem> items, int estimatedTotal) {}
